You can now add RegularUsers when you are logged in.
This has started to cause some issues in for example the MyMenuActivity, as we do not want Regular Users to add Coworkers.
We just want to check the boolean variable saved under RegularUsers information and (soon) BossUsers information.
true userType = Regular
false userType = Boss.
And if true, we will need to start another MyMenuActivity (Not made yet) and when trying to access his User information.
We want to approach the UserInfoActivity a bit different because the Regular Users Information is stored a little different than the Boss Users' Information.
This is also why I am saving the bossID inside the Regular Users, because when we log in as a regular User. we want to be able to access some of the Bosses data.
Because all of the Regular Users data is stored inside the bossID, RegularUsers and then under his own userID.

I have chosen to save the Data like this because then someone can't accidentally access someone elses data, because they will always access the data under his userID. So unless this is changed (Possible, but also possible to prevent) the data will be secure from outsiders without the users email and password.