The MyMenu for the Regular Users is made.
Additional changes where made in almost every class, as I changed the dropdownlisteners Strign Array to be Danish. The Listener for that had to be updated.
I will change the MyMenu for Boss users tomorrow. So the boss can see the contracts button. And then I will change the WorkersActivity so that there is an AddWorkers button in there instead.
And I will make it so the Regular Users can see their Coworkers and themselves. This is done because at one point the Users wan't to add other users to Tasks. So If I do it like this, the method for seeing his coworkers is already done. And in the future, it will be easier to implement.