The MyMenu and MyMenu2 is setup now. It only needs to start up the contracts activity and the tasks activity.
When the app starts it checks if the user is authenticated, if so, it checks the userType. if RegularUser is true, it opens MyMenu2. If false, it opens MyMenu.
(MyMenu and MyMenu2 is exactly the same right now. It is explained later why.)

Then if the user clicks the "Medarbejdere" button, it check the users UserType again onCreate.
If regularUser is true, it uses a different Database reference, as it needs to check under the bossID stored in his own data.
if false, it uses the UserID databsae reference, because if it is a boss that is currently signed in, the Coworkers is stored under his own UserID.

This is why the MyMenu & MyMenu2 is the same. It checks the user when it open up the other activities. So the Regular Users can still see their Coworkers but they cannot see the addWorkerBtn. This is done because later we might want to add workers to the tasks. If so, the Regular Users must be able to see the workers as well as the boss.