Changes made all over the System.
I have implemented Customers in the system. So whenever you create an assignment. Thw details of the worker you have entered will be compared to your previous Customers, if none matches, then it creates a new Customer. However if the Customer already Exists in your data, aka is a previous Customer, the Assignment/Task that is created, that Task's taskID is stored under the Customer. So every Customer has at least one Task linked to him, if there is several, several is linked to.

I have also made the change so that the CheckUserType boolean and the bossID is send through every Activty that is created, this is needed in mostly all Activities now, because they all depend on what type of user is logged in.

When starting an Activty use following code

Intent intent = new Intent();
intent.putExtras("userType", userType);
intent.putExtras("bossID", bossID);
startActivity(intent);

and in the onCreate method in the new activty, if these values is needed. (This code is included in almost every Activity Already. Only ContractActivty needs this I believe)
Intent intent = this.getIntet();
userType = intent.getExtras().getBoolean("userType");
bossID = intent.getExtras().getString("bossID");

(This might be wrong, if it is, just check the start of one of the onCreate methods in for example MyMenuActivity.)